requirements.txt
absl-py==0.11.0
appdirs==1.4.4
asgiref==3.3.1
astunparse==1.6.3
beautifulsoup4==4.9.3
Brotli==1.0.9
cachelib==0.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
cryptography==3.3.1
cssselect==1.1.0
defusedxml==0.6.0
Deprecated==1.2.10
dj-database-url==0.5.0
Django==3.1.5
django-allauth==0.44.0
django-classy-tags==2.0.0
django-cms==3.8.0
django-formtools==2.2
django-heroku==0.3.1
django-sekizai==2.0.0
django-treebeard==4.4
djangocms-admin-style==2.0.2
flatbuffers==1.12
gast==0.3.3
gcloud==0.18.3
google==3.0.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
google-trans-new==1.1.9
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
h5py==2.10.0
howdoi==2.0.8
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httplib2==0.18.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.0.0
jwcrypto==0.8
keep==2.10
Keras-Preprocessing==1.1.2
lxml==4.6.2
Markdown==3.3.3
more-itertools==8.6.0
nltk==3.5
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
praw==7.1.0
prawcore==1.5.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
PyGithub==1.54.1
Pygments==2.7.4
pyjokes==0.6.0
PyJWT==1.7.1
pyquery==1.4.3
Pyrebase4==4.4.3
python-jwt==3.3.0
python3-openid==3.2.0
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.7
scikit-learn==0.24.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==2.1.0
tqdm==4.56.0
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
Werkzeug==1.0.1
whitenoise==5.2.0
wikipedia==1.4.0
wolframalpha==4.1.1
wrapt==1.12.1
xmltodict==0.12.0
